requirements.txt
annotated-types==0.7.0
anyio==3.7.1
audioread==3.0.1
certifi==2025.10.5
charset-normalizer==3.4.4
dnspython==2.8.0
fastapi==0.104.1
google-genai==0.8.0
h11==0.16.0
httptools==0.7.1
idna==3.11
librosa==0.11.0
motor==3.3.2
numpy==2.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pymongo==4.6.0
python-dotenv==1.0.0
python-multipart==0.0.20
regex==2025.10.23
sniffio==1.3.1
soundfile==0.13.1
starlette==0.27.0
transformers==4.57.1
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
